gran avance, falta lang

+++ src/main/java/com/herobrot/tieredneo/network/payload/ReforgeScreenPayload.java
@@ -7,12 +7,12 @@ import net.minecraft.network.codec.StreamCodec;
 import net.minecraft.network.protocol.common.custom.CustomPacketPayload;
 import org.jetbrains.annotations.NotNull;
 
-public record ReforgeScreenPayload(int mouseX, int mouseY, boolean reforgingScreen) implements CustomPacketPayload {
+public record ReforgeScreenPayload(double mouseX, double mouseY, boolean reforgingScreen) implements CustomPacketPayload {
     public static final Type<ReforgeScreenPayload> TYPE = new Type<>(TieredNeo.rl("reforge_screen"));
 
     public static final StreamCodec<ByteBuf, ReforgeScreenPayload> STREAM_CODEC = StreamCodec.composite(
-            ByteBufCodecs.INT, ReforgeScreenPayload::mouseX,
-            ByteBufCodecs.INT, ReforgeScreenPayload::mouseY,
+            ByteBufCodecs.DOUBLE, ReforgeScreenPayload::mouseX,
+            ByteBufCodecs.DOUBLE, ReforgeScreenPayload::mouseY,
             ByteBufCodecs.BOOL, ReforgeScreenPayload::reforgingScreen,
             ReforgeScreenPayload::new
     );
